refactor code, redesign app
