+++ app/src/main/java/com/loftymr/whichone/feature/component/CircularProgressAnimated.kt
@@ -1,0 +1,45 @@
+package com.loftymr.whichone.feature.component
+
+import androidx.compose.animation.core.animateFloat
+import androidx.compose.animation.core.infiniteRepeatable
+import androidx.compose.animation.core.rememberInfiniteTransition
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+
+/**
+ * Created by talhafaki on 13.09.2022.
+ */
+
+@Composable
+fun CircularProgressAnimated(modifier: Modifier = Modifier) {
+    val progressValue = 1f
+    val infiniteTransition = rememberInfiniteTransition()
+
+    val progressAnimationValue by infiniteTransition.animateFloat(
+        initialValue = 0.0f,
+        targetValue = progressValue, animationSpec = infiniteRepeatable(animation = tween(900))
+    )
+
+    Box(
+        modifier = modifier
+            .fillMaxSize()
+            .background(color = Color(0xFF3552A2)),
+        contentAlignment = Alignment.Center
+    ) {
+        CircularProgressIndicator(
+            progress = progressAnimationValue,
+            modifier = Modifier
+                .wrapContentSize(),
+            color = Color.White.copy(alpha = 0.8f)
+        )
+    }
+}
