+++ app/src/main/java/com/loftymr/whichone/feature/component/StepsProgressBar.kt
@@ -1,0 +1,64 @@
+package com.loftymr.whichone.feature.component
+
+import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.Divider
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+/**
+ * Created by talhafaki on 13.09.2022.
+ */
+
+@Composable
+fun StepsProgressBar(modifier: Modifier = Modifier, numberOfSteps: Int, currentStep: Int) {
+    Row(
+        modifier = modifier,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        for (step in 0..numberOfSteps) {
+            Step(
+                modifier = Modifier.weight(1F),
+                isCompete = step < currentStep,
+                isCurrent = step == currentStep
+            )
+        }
+    }
+}
+
+@Composable
+fun Step(modifier: Modifier = Modifier, isCompete: Boolean, isCurrent: Boolean) {
+    val color = if (isCompete || isCurrent) Color(0xFF3552A2) else Color.LightGray
+    val innerCircleColor = if (isCompete) Color(0xFF3552A2) else Color.LightGray
+
+    Box(modifier = modifier) {
+
+        //Line
+        Divider(
+            modifier = Modifier.align(Alignment.CenterStart),
+            color = color,
+            thickness = 2.dp
+        )
+
+        //Circle
+        Canvas(modifier = Modifier
+            .size(15.dp)
+            .align(Alignment.CenterEnd)
+            .border(
+                shape = CircleShape,
+                width = 2.dp,
+                color = color
+            ),
+            onDraw = {
+                drawCircle(color = innerCircleColor)
+            }
+        )
+    }
+}
