+++ app/src/main/java/com/loftymr/whichone/feature/screen/main/MainActivity.kt
@@ -5,13 +5,10 @@ import android.net.Uri
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
 import com.google.android.gms.ads.MobileAds
 import com.google.firebase.FirebaseApp
 import com.loftymr.whichone.feature.navigation.NavGraph
@@ -35,20 +32,12 @@ class MainActivity : ComponentActivity() {
 
             }).check()
         MobileAds.initialize(this)
-        val verticalGradientBrush = Brush.verticalGradient(
-            colors = listOf(
-                Color(0xFFC8D7CF),
-                Color(0xFF6987B4),
-                Color(0xFF333D5F)
-            )
-        )
         setContent {
             WhichOneTheme(darkTheme = false) {
                 // A surface container using the 'background' color from the theme
                 Surface(
                     modifier = Modifier
-                        .fillMaxSize()
-                        .background(brush = verticalGradientBrush),
+                        .fillMaxSize(),
                     color = MaterialTheme.colors.background
                 ) {
                     NavGraph(
