+++ app/src/main/java/com/loftymr/whichone/feature/screen/result/ResultScreen.kt
@@ -10,13 +10,16 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.rememberImagePainter
 import com.loftymr.whichone.feature.component.WhichOneButton
 import com.loftymr.whichone.feature.component.WhichOneTemplate
+import com.loftymr.whichone.R
 
 /**
  * Created by talhafaki on 10.09.2022.
@@ -62,6 +65,7 @@ fun ResultContent(
             modifier = Modifier
                 .fillMaxWidth()
                 .padding(16.dp)
+                .shadow(elevation = 10.dp, shape = RoundedCornerShape(8.dp))
                 .clip(RoundedCornerShape(8.dp))
         )
 
@@ -71,7 +75,7 @@ fun ResultContent(
                 .fillMaxWidth()
                 .wrapContentWidth(Alignment.CenterHorizontally)
                 .padding(horizontal = 16.dp, vertical = 8.dp),
-            color = Color.DarkGray,
+            color = Color.White,
             style = MaterialTheme.typography.body1.copy(color = Color.White, fontSize = 24.sp)
         )
 
@@ -82,7 +86,7 @@ fun ResultContent(
                 .padding(bottom = 24.dp),
             contentAlignment = Alignment.BottomCenter
         ) {
-            WhichOneButton(buttonText = "AGAIN") {
+            WhichOneButton(buttonText = stringResource(id = R.string.again)) {
                 navigateToSurvey.invoke()
             }
         }
