+++ app/src/main/java/com/loftymr/whichone/feature/theme/Color.kt
@@ -6,3 +6,5 @@ val Purple200 = Color(0xFFBB86FC)
 val Purple500 = Color(0xFF6200EE)
 val Purple700 = Color(0xFF3700B3)
 val Teal200 = Color(0xFF03DAC5)
+
+val SolidBlue = Color(0xFF3552A2)
