+++ app/src/main/java/com/loftymr/whichone/util/extension/Modifier.kt
@@ -1,0 +1,18 @@
+package com.loftymr.whichone.util.extension
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.composed
+
+/**
+ * Created by talhafaki on 13.09.2022.
+ */
+
+inline fun Modifier.noRippleClickable(crossinline onClick: ()->Unit): Modifier = composed {
+    clickable(indication = null,
+        interactionSource = remember { MutableInteractionSource() }) {
+        onClick()
+    }
+}
